+++ packages/natives/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+### Changed
+- Allowed PTY sessions to accept shell preference and adjust command invocation for cmd and PowerShell
 ## [12.1.0] - 2026-02-13
 ### Added